Updated SimplePie > 1.8.1
