+++ src/lib/SimplePie/src/Cache/BaseDataCache.php
@@ -1,15 +1,9 @@
 <?php
 
+// SPDX-FileCopyrightText: 2004-2023 Ryan Parman, Sam Sneddon, Ryan McCue
+// SPDX-License-Identifier: BSD-3-Clause
+
 declare(strict_types=1);
-/**
- * @package SimplePie
- * @copyright 2004-2022 Ryan Parman, Sam Sneddon, Ryan McCue
- * @author Ryan Parman
- * @author Sam Sneddon
- * @author Ryan McCue
- * @link http://simplepie.org/ SimplePie
- * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- */
 
 namespace SimplePie\Cache;
 
@@ -18,8 +12,6 @@ use InvalidArgumentException;
 /**
  * Adapter for deprecated \SimplePie\Cache\Base implementations
  *
- * @package SimplePie
- * @subpackage Caching
  * @internal
  */
 final class BaseDataCache implements DataCache
@@ -54,12 +46,12 @@ final class BaseDataCache implements DataCache
     {
         $data = $this->cache->load();
 
-        if (! is_array($data)) {
+        if (!is_array($data)) {
             return $default;
         }
 
         // ignore data if internal cache expiration time is not set
-        if (! array_key_exists('__cache_expiration_time', $data)) {
+        if (!array_key_exists('__cache_expiration_time', $data)) {
             return $default;
         }
 
@@ -83,7 +75,7 @@ final class BaseDataCache implements DataCache
      * </code>
      *
      * @param string   $key   The key of the item to store.
-     * @param array    $value The value of the item to store, must be serializable.
+     * @param array<mixed> $value The value of the item to store, must be serializable.
      * @param null|int $ttl   Optional. The TTL value of this item. If no value is sent and
      *                                      the driver supports TTL then the library may set a default value
      *                                      for it or let the driver take care of that.
