+++ src/lib/SimplePie/src/Cache/Memcached.php
@@ -1,15 +1,10 @@
 <?php
 
+// SPDX-FileCopyrightText: 2004-2023 Ryan Parman, Sam Sneddon, Ryan McCue
+// SPDX-FileCopyrightText: 2015 Paul L. McNeely
+// SPDX-License-Identifier: BSD-3-Clause
+
 declare(strict_types=1);
-/**
- * @package SimplePie
- * @copyright 2004-2016 Ryan Parman, Sam Sneddon, Ryan McCue
- * @author Ryan Parman
- * @author Sam Sneddon
- * @author Ryan McCue
- * @link http://simplepie.org/ SimplePie
- * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- */
 
 namespace SimplePie\Cache;
 
@@ -24,9 +19,6 @@ use Memcached as NativeMemcached;
  * connect to memcached on `localhost` on port 11211. All tables will be
  * prefixed with `sp_` and data will expire after 3600 seconds
  *
- * @package    SimplePie
- * @subpackage Caching
- * @author     Paul L. McNeely
  * @uses       Memcached
  * @deprecated since SimplePie 1.8.0, use implementation of "Psr\SimpleCache\CacheInterface" instead
  */
@@ -40,7 +32,7 @@ class Memcached implements Base
 
     /**
      * Options
-     * @var array
+     * @var array<string, mixed>
      */
     protected $options;
 
@@ -56,7 +48,7 @@ class Memcached implements Base
      * @param string $name Unique ID for the cache
      * @param Base::TYPE_FEED|Base::TYPE_IMAGE $type Either TYPE_FEED for SimplePie data, or TYPE_IMAGE for image data
      */
-    public function __construct($location, $name, $type)
+    public function __construct(string $location, string $name, $type)
     {
         $this->options = [
             'host'   => '127.0.0.1',
@@ -76,7 +68,7 @@ class Memcached implements Base
 
     /**
      * Save data to the cache
-     * @param array|\SimplePie\SimplePie $data Data to store in the cache. If passed a SimplePie object, only cache the $data property
+     * @param array<mixed>|\SimplePie\SimplePie $data Data to store in the cache. If passed a SimplePie object, only cache the $data property
      * @return bool Successfulness
      */
     public function save($data)
@@ -90,7 +82,7 @@ class Memcached implements Base
 
     /**
      * Retrieve the data saved to the cache
-     * @return array Data for SimplePie::$data
+     * @return array<mixed>|false Data for SimplePie::$data
      */
     public function load()
     {
@@ -133,9 +125,10 @@ class Memcached implements Base
 
     /**
      * Set the last modified time and data to NativeMemcached
+     * @param string|false $data
      * @return bool Success status
      */
-    private function setData($data)
+    private function setData($data): bool
     {
         if ($data !== false) {
             $this->cache->set($this->name . '_mtime', time(), (int)$this->options['extras']['timeout']);
