+++ src/lib/SimplePie/src/Content/Type/Sniffer.php
@@ -1,17 +1,15 @@
 <?php
 
+// SPDX-FileCopyrightText: 2004-2023 Ryan Parman, Sam Sneddon, Ryan McCue
+// SPDX-License-Identifier: BSD-3-Clause
+
 declare(strict_types=1);
-/**
- * @package SimplePie
- * @copyright 2004-2016 Ryan Parman, Sam Sneddon, Ryan McCue
- * @author Ryan Parman
- * @author Sam Sneddon
- * @author Ryan McCue
- * @link http://simplepie.org/ SimplePie
- * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- */
 
 namespace SimplePie\Content\Type;
+
+use InvalidArgumentException;
+use SimplePie\File;
+use SimplePie\HTTP\Response;
 
 /**
  * Content-type sniffing
@@ -23,26 +21,32 @@ namespace SimplePie\Content\Type;
  *
  *
  * This class can be overloaded with {@see \SimplePie\SimplePie::set_content_type_sniffer_class()}
- *
- * @package SimplePie
- * @subpackage HTTP
  */
 class Sniffer
 {
     /**
      * File object
      *
-     * @var \SimplePie\File
+     * @var File|Response
      */
     public $file;
 
     /**
      * Create an instance of the class with the input file
      *
-     * @param Sniffer $file Input file
+     * @param File|Response $file Input file
      */
-    public function __construct($file)
+    public function __construct(/* File */ $file)
     {
+        if (!is_object($file) || !$file instanceof Response) {
+            // For BC we're asking for `File`, but internally we accept every `Response` implementation
+            throw new InvalidArgumentException(sprintf(
+                '%s(): Argument #1 ($file) must be of type %s',
+                __METHOD__,
+                File::class
+            ), 1);
+        }
+
         $this->file = $file;
     }
 
@@ -53,19 +57,21 @@ class Sniffer
      */
     public function get_type()
     {
-        if (isset($this->file->headers['content-type'])) {
-            if (!isset($this->file->headers['content-encoding'])
-                && ($this->file->headers['content-type'] === 'text/plain'
-                    || $this->file->headers['content-type'] === 'text/plain; charset=ISO-8859-1'
-                    || $this->file->headers['content-type'] === 'text/plain; charset=iso-8859-1'
-                    || $this->file->headers['content-type'] === 'text/plain; charset=UTF-8')) {
+        $content_type = $this->file->has_header('content-type') ? $this->file->get_header_line('content-type') : null;
+        $content_encoding = $this->file->has_header('content-encoding') ? $this->file->get_header_line('content-encoding') : null;
+        if ($content_type !== null) {
+            if ($content_encoding === null
+                && ($content_type === 'text/plain'
+                    || $content_type === 'text/plain; charset=ISO-8859-1'
+                    || $content_type === 'text/plain; charset=iso-8859-1'
+                    || $content_type === 'text/plain; charset=UTF-8')) {
                 return $this->text_or_binary();
             }
 
-            if (($pos = strpos($this->file->headers['content-type'], ';')) !== false) {
-                $official = substr($this->file->headers['content-type'], 0, $pos);
+            if (($pos = strpos($content_type, ';')) !== false) {
+                $official = substr($content_type, 0, $pos);
             } else {
-                $official = $this->file->headers['content-type'];
+                $official = $content_type;
             }
             $official = trim(strtolower($official));
 
@@ -99,12 +105,14 @@ class Sniffer
      */
     public function text_or_binary()
     {
-        if (substr($this->file->body, 0, 2) === "\xFE\xFF"
-            || substr($this->file->body, 0, 2) === "\xFF\xFE"
-            || substr($this->file->body, 0, 4) === "\x00\x00\xFE\xFF"
-            || substr($this->file->body, 0, 3) === "\xEF\xBB\xBF") {
+        $body = $this->file->get_body_content();
+
+        if (substr($body, 0, 2) === "\xFE\xFF"
+            || substr($body, 0, 2) === "\xFF\xFE"
+            || substr($body, 0, 4) === "\x00\x00\xFE\xFF"
+            || substr($body, 0, 3) === "\xEF\xBB\xBF") {
             return 'text/plain';
-        } elseif (preg_match('/[\x00-\x08\x0E-\x1A\x1C-\x1F]/', $this->file->body)) {
+        } elseif (preg_match('/[\x00-\x08\x0E-\x1A\x1C-\x1F]/', $body)) {
             return 'application/octet-stream';
         }
 
@@ -118,25 +126,27 @@ class Sniffer
      */
     public function unknown()
     {
-        $ws = strspn($this->file->body, "\x09\x0A\x0B\x0C\x0D\x20");
-        if (strtolower(substr($this->file->body, $ws, 14)) === '<!doctype html'
-            || strtolower(substr($this->file->body, $ws, 5)) === '<html'
-            || strtolower(substr($this->file->body, $ws, 7)) === '<script') {
+        $body = $this->file->get_body_content();
+
+        $ws = strspn($body, "\x09\x0A\x0B\x0C\x0D\x20");
+        if (strtolower(substr($body, $ws, 14)) === '<!doctype html'
+            || strtolower(substr($body, $ws, 5)) === '<html'
+            || strtolower(substr($body, $ws, 7)) === '<script') {
             return 'text/html';
-        } elseif (substr($this->file->body, 0, 5) === '%PDF-') {
+        } elseif (substr($body, 0, 5) === '%PDF-') {
             return 'application/pdf';
-        } elseif (substr($this->file->body, 0, 11) === '%!PS-Adobe-') {
+        } elseif (substr($body, 0, 11) === '%!PS-Adobe-') {
             return 'application/postscript';
-        } elseif (substr($this->file->body, 0, 6) === 'GIF87a'
-            || substr($this->file->body, 0, 6) === 'GIF89a') {
+        } elseif (substr($body, 0, 6) === 'GIF87a'
+            || substr($body, 0, 6) === 'GIF89a') {
             return 'image/gif';
-        } elseif (substr($this->file->body, 0, 8) === "\x89\x50\x4E\x47\x0D\x0A\x1A\x0A") {
+        } elseif (substr($body, 0, 8) === "\x89\x50\x4E\x47\x0D\x0A\x1A\x0A") {
             return 'image/png';
-        } elseif (substr($this->file->body, 0, 3) === "\xFF\xD8\xFF") {
+        } elseif (substr($body, 0, 3) === "\xFF\xD8\xFF") {
             return 'image/jpeg';
-        } elseif (substr($this->file->body, 0, 2) === "\x42\x4D") {
+        } elseif (substr($body, 0, 2) === "\x42\x4D") {
             return 'image/bmp';
-        } elseif (substr($this->file->body, 0, 4) === "\x00\x00\x01\x00") {
+        } elseif (substr($body, 0, 4) === "\x00\x00\x01\x00") {
             return 'image/vnd.microsoft.icon';
         }
 
@@ -146,20 +156,22 @@ class Sniffer
     /**
      * Sniff images
      *
-     * @return string Actual Content-Type
+     * @return string|false Actual Content-Type
      */
     public function image()
     {
-        if (substr($this->file->body, 0, 6) === 'GIF87a'
-            || substr($this->file->body, 0, 6) === 'GIF89a') {
+        $body = $this->file->get_body_content();
+
+        if (substr($body, 0, 6) === 'GIF87a'
+            || substr($body, 0, 6) === 'GIF89a') {
             return 'image/gif';
-        } elseif (substr($this->file->body, 0, 8) === "\x89\x50\x4E\x47\x0D\x0A\x1A\x0A") {
+        } elseif (substr($body, 0, 8) === "\x89\x50\x4E\x47\x0D\x0A\x1A\x0A") {
             return 'image/png';
-        } elseif (substr($this->file->body, 0, 3) === "\xFF\xD8\xFF") {
+        } elseif (substr($body, 0, 3) === "\xFF\xD8\xFF") {
             return 'image/jpeg';
-        } elseif (substr($this->file->body, 0, 2) === "\x42\x4D") {
+        } elseif (substr($body, 0, 2) === "\x42\x4D") {
             return 'image/bmp';
-        } elseif (substr($this->file->body, 0, 4) === "\x00\x00\x01\x00") {
+        } elseif (substr($body, 0, 4) === "\x00\x00\x01\x00") {
             return 'image/vnd.microsoft.icon';
         }
 
@@ -173,16 +185,18 @@ class Sniffer
      */
     public function feed_or_html()
     {
-        $len = strlen($this->file->body);
-        $pos = strspn($this->file->body, "\x09\x0A\x0D\x20\xEF\xBB\xBF");
+        $body = $this->file->get_body_content();
+
+        $len = strlen($body);
+        $pos = strspn($body, "\x09\x0A\x0D\x20\xEF\xBB\xBF");
 
         while ($pos < $len) {
-            switch ($this->file->body[$pos]) {
+            switch ($body[$pos]) {
                 case "\x09":
                 case "\x0A":
                 case "\x0D":
                 case "\x20":
-                    $pos += strspn($this->file->body, "\x09\x0A\x0D\x20", $pos);
+                    $pos += strspn($body, "\x09\x0A\x0D\x20", $pos);
                     continue 2;
 
                 case '<':
@@ -193,29 +207,29 @@ class Sniffer
                     return 'text/html';
             }
 
-            if (substr($this->file->body, $pos, 3) === '!--') {
+            if (substr($body, $pos, 3) === '!--') {
                 $pos += 3;
-                if ($pos < $len && ($pos = strpos($this->file->body, '-->', $pos)) !== false) {
+                if ($pos < $len && ($pos = strpos($body, '-->', $pos)) !== false) {
                     $pos += 3;
                 } else {
                     return 'text/html';
                 }
-            } elseif (substr($this->file->body, $pos, 1) === '!') {
-                if ($pos < $len && ($pos = strpos($this->file->body, '>', $pos)) !== false) {
+            } elseif (substr($body, $pos, 1) === '!') {
+                if ($pos < $len && ($pos = strpos($body, '>', $pos)) !== false) {
                     $pos++;
                 } else {
                     return 'text/html';
                 }
-            } elseif (substr($this->file->body, $pos, 1) === '?') {
-                if ($pos < $len && ($pos = strpos($this->file->body, '?>', $pos)) !== false) {
+            } elseif (substr($body, $pos, 1) === '?') {
+                if ($pos < $len && ($pos = strpos($body, '?>', $pos)) !== false) {
                     $pos += 2;
                 } else {
                     return 'text/html';
                 }
-            } elseif (substr($this->file->body, $pos, 3) === 'rss'
-                || substr($this->file->body, $pos, 7) === 'rdf:RDF') {
+            } elseif (substr($body, $pos, 3) === 'rss'
+                || substr($body, $pos, 7) === 'rdf:RDF') {
                 return 'application/rss+xml';
-            } elseif (substr($this->file->body, $pos, 4) === 'feed') {
+            } elseif (substr($body, $pos, 4) === 'feed') {
                 return 'application/atom+xml';
             } else {
                 return 'text/html';
