+++ src/lib/SimplePie/src/HTTP/Parser.php
@@ -1,23 +1,15 @@
 <?php
 
+// SPDX-FileCopyrightText: 2004-2023 Ryan Parman, Sam Sneddon, Ryan McCue
+// SPDX-License-Identifier: BSD-3-Clause
+
 declare(strict_types=1);
-/**
- * @package SimplePie
- * @copyright 2004-2016 Ryan Parman, Sam Sneddon, Ryan McCue
- * @author Ryan Parman
- * @author Sam Sneddon
- * @author Ryan McCue
- * @link http://simplepie.org/ SimplePie
- * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- */
 
 namespace SimplePie\HTTP;
 
 /**
  * HTTP Response Parser
- *
- * @package SimplePie
- * @subpackage HTTP
+ * @template Psr7Compatible of bool
  */
 class Parser
 {
@@ -43,9 +35,14 @@ class Parser
     public $reason = '';
 
     /**
+     * @var Psr7Compatible whether headers are compatible with PSR-7 format.
+     */
+    private $psr7Compatible;
+
+    /**
      * Key/value pairs of the headers
      *
-     * @var array
+     * @var (Psr7Compatible is true ? array<string, non-empty-array<string>> : array<string, string>)
      */
     public $headers = [];
 
@@ -113,14 +110,14 @@ class Parser
     protected $position = 0;
 
     /**
-     * Name of the hedaer currently being parsed
+     * Name of the header currently being parsed
      *
      * @var string
      */
     protected $name = '';
 
     /**
-     * Value of the hedaer currently being parsed
+     * Value of the header currently being parsed
      *
      * @var string
      */
@@ -130,11 +127,13 @@ class Parser
      * Create an instance of the class with the input data
      *
      * @param string $data Input data
+     * @param Psr7Compatible $psr7Compatible Whether the data types are in format compatible with PSR-7.
      */
-    public function __construct($data)
+    public function __construct(string $data, bool $psr7Compatible = false)
     {
         $this->data = $data;
         $this->data_length = strlen($this->data);
+        $this->psr7Compatible = $psr7Compatible;
     }
 
     /**
@@ -153,7 +152,8 @@ class Parser
             return true;
         }
 
-        $this->http_version = '';
+        // Reset the parser state.
+        $this->http_version = 0.0;
         $this->status_code = 0;
         $this->reason = '';
         $this->headers = [];
@@ -187,15 +187,16 @@ class Parser
 
     /**
      * Parse the HTTP version
+     * @return void
      */
     protected function http_version()
     {
         if (strpos($this->data, "\x0A") !== false && strtoupper(substr($this->data, 0, 5)) === 'HTTP/') {
             $len = strspn($this->data, '0123456789.', 5);
-            $this->http_version = substr($this->data, 5, $len);
+            $http_version = substr($this->data, 5, $len);
             $this->position += 5 + $len;
-            if (substr_count($this->http_version, '.') <= 1) {
-                $this->http_version = (float) $this->http_version;
+            if (substr_count($http_version, '.') <= 1) {
+                $this->http_version = (float) $http_version;
                 $this->position += strspn($this->data, "\x09\x20", $this->position);
                 $this->state = self::STATE_STATUS;
             } else {
@@ -208,6 +209,7 @@ class Parser
 
     /**
      * Parse the status code
+     * @return void
      */
     protected function status()
     {
@@ -222,6 +224,7 @@ class Parser
 
     /**
      * Parse the reason phrase
+     * @return void
      */
     protected function reason()
     {
@@ -233,6 +236,7 @@ class Parser
 
     /**
      * Deal with a new line, shifting data around as needed
+     * @return void
      */
     protected function new_line()
     {
@@ -241,9 +245,17 @@ class Parser
             $this->name = strtolower($this->name);
             // We should only use the last Content-Type header. c.f. issue #1
             if (isset($this->headers[$this->name]) && $this->name !== 'content-type') {
-                $this->headers[$this->name] .= ', ' . $this->value;
+                if ($this->psr7Compatible) {
+                    $this->headers[$this->name][] = $this->value;
+                } else {
+                    $this->headers[$this->name] .= ', ' . $this->value;
+                }
             } else {
-                $this->headers[$this->name] = $this->value;
+                if ($this->psr7Compatible) {
+                    $this->headers[$this->name] = [$this->value];
+                } else {
+                    $this->headers[$this->name] = $this->value;
+                }
             }
         }
         $this->name = '';
@@ -261,6 +273,7 @@ class Parser
 
     /**
      * Parse a header name
+     * @return void
      */
     protected function name()
     {
@@ -281,6 +294,7 @@ class Parser
 
     /**
      * Parse LWS, replacing consecutive LWS characters with a single space
+     * @return void
      */
     protected function linear_whitespace()
     {
@@ -297,6 +311,7 @@ class Parser
 
     /**
      * See what state to move to while within non-quoted header values
+     * @return void
      */
     protected function value()
     {
@@ -331,6 +346,7 @@ class Parser
 
     /**
      * Parse a header value while outside quotes
+     * @return void
      */
     protected function value_char()
     {
@@ -342,6 +358,7 @@ class Parser
 
     /**
      * See what state to move to while within quoted header values
+     * @return void
      */
     protected function quote()
     {
@@ -373,6 +390,7 @@ class Parser
 
     /**
      * Parse a header value while within quotes
+     * @return void
      */
     protected function quote_char()
     {
@@ -384,6 +402,7 @@ class Parser
 
     /**
      * Parse an escaped character within quotes
+     * @return void
      */
     protected function quote_escaped()
     {
@@ -394,6 +413,7 @@ class Parser
 
     /**
      * Parse the body
+     * @return void
      */
     protected function body()
     {
@@ -408,6 +428,7 @@ class Parser
 
     /**
      * Parsed a "Transfer-Encoding: chunked" body
+     * @return void
      */
     protected function chunked()
     {
@@ -454,11 +475,11 @@ class Parser
      * Prepare headers (take care of proxies headers)
      *
      * @param string  $headers Raw headers
-     * @param integer $count   Redirection count. Default to 1.
+     * @param int $count   Redirection count. Default to 1.
      *
      * @return string
      */
-    public static function prepareHeaders($headers, $count = 1)
+    public static function prepareHeaders(string $headers, int $count = 1)
     {
         $data = explode("\r\n\r\n", $headers, $count);
         $data = array_pop($data);
