+++ src/lib/SimplePie/src/Net/IPv6.php
@@ -1,23 +1,15 @@
 <?php
 
+// SPDX-FileCopyrightText: 2004-2023 Ryan Parman, Sam Sneddon, Ryan McCue
+// SPDX-License-Identifier: BSD-3-Clause
+
 declare(strict_types=1);
-/**
- * @package SimplePie
- * @copyright 2004-2016 Ryan Parman, Sam Sneddon, Ryan McCue
- * @author Ryan Parman
- * @author Sam Sneddon
- * @author Ryan McCue
- * @link http://simplepie.org/ SimplePie
- * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- */
 
 namespace SimplePie\Net;
 
 /**
  * Class to validate and to work with IPv6 addresses.
  *
- * @package SimplePie
- * @subpackage HTTP
  * @copyright 2003-2005 The PHP Group
  * @license http://www.opensource.org/licenses/bsd-license.php
  * @link http://pear.php.net/package/Net_IPv6
@@ -46,7 +38,7 @@ class IPv6
      * @param string $ip An IPv6 address
      * @return string The uncompressed IPv6 address
      */
-    public static function uncompress($ip)
+    public static function uncompress(string $ip)
     {
         $c1 = -1;
         $c2 = -1;
@@ -102,7 +94,7 @@ class IPv6
      * @param string $ip An IPv6 address
      * @return string The compressed IPv6 address
      */
-    public static function compress($ip)
+    public static function compress(string $ip)
     {
         // Prepare the IP to be compressed
         $ip = self::uncompress($ip);
@@ -142,9 +134,9 @@ class IPv6
      *           0:0:0:0:0:FFFF:129.144.52.38
      *
      * @param string $ip An IPv6 address
-     * @return array [0] contains the IPv6 represented part, and [1] the IPv4 represented part
+     * @return array{string, string} [0] contains the IPv6 represented part, and [1] the IPv4 represented part
      */
-    private static function split_v6_v4($ip)
+    private static function split_v6_v4(string $ip): array
     {
         if (strpos($ip, '.') !== false) {
             $pos = strrpos($ip, ':');
@@ -164,7 +156,7 @@ class IPv6
      * @param string $ip An IPv6 address
      * @return bool true if $ip is a valid IPv6 address
      */
-    public static function check_ipv6($ip)
+    public static function check_ipv6(string $ip)
     {
         $ip = self::uncompress($ip);
         [$ipv6, $ipv4] = self::split_v6_v4($ip);
@@ -217,7 +209,7 @@ class IPv6
      * @param string $ip An IPv6 address
      * @return bool true if $ip is a valid IPv6 address
      */
-    public static function checkIPv6($ip)
+    public static function checkIPv6(string $ip)
     {
         return self::check_ipv6($ip);
     }
